+++ src/main/java/dandi/dandi/member/application/service/MemberQueryServiceAdapter.java
@@ -5,8 +5,8 @@ import dandi.dandi.member.application.port.in.MemberInfoResponse;
 import dandi.dandi.member.application.port.in.MemberQueryServicePort;
 import dandi.dandi.member.application.port.in.NicknameDuplicationCheckResponse;
 import dandi.dandi.member.application.port.out.MemberPersistencePort;
+import dandi.dandi.member.application.port.out.MemberPostPersistencePort;
 import dandi.dandi.member.domain.Member;
-import dandi.dandi.post.application.port.out.MemberPostPersistencePort;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
