+++ src/test/java/dandi/dandi/member/application/service/MemberQueryServiceAdapterTest.java
@@ -13,8 +13,8 @@ import dandi.dandi.auth.exception.UnauthorizedException;
 import dandi.dandi.member.application.port.in.MemberInfoResponse;
 import dandi.dandi.member.application.port.in.NicknameDuplicationCheckResponse;
 import dandi.dandi.member.application.port.out.MemberPersistencePort;
+import dandi.dandi.member.application.port.out.MemberPostPersistencePort;
 import dandi.dandi.member.domain.Member;
-import dandi.dandi.post.application.port.out.MemberPostPersistencePort;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
